requirements.txt
# Data handling and analysis
pandas==1.4.4
numpy==1.23.3

# Machine Learning
torch==1.12.1
torchvision==0.13.1

# Image processing
Pillow==9.2.0

# Visualization
matplotlib==3.5.3
seaborn==0.12.0

# Web application
streamlit==1.37.1

# Model explainability
lime==0.2.0.1
shap==0.41.0

# Evaluation metrics
scikit-learn==1.1.2

# Notebook environment
jupyter==1.0.0
ipython==8.5.0

# Progress bars
tqdm==4.64.1

# For handling DICOM images (if needed)
pydicom==2.3.0

# For advanced data augmentation (if needed)
albumentations==1.3.0

# For hyperparameter tuning (if needed)
optuna==3.0.3

# For experiment tracking (if needed)
mlflow==1.29.0

# For deep learning model architectures (if needed)
efficientnet-pytorch==0.7.1

# For API creation (if needed for deployment)
fastapi==0.85.0
uvicorn==0.18.3

